Added Pin Translation
This allows chips where logical pins don't line up with analog pins. For
example, RA2 is AN4 on the PIC24F16KA101.

I also added in some catch code to ensure the A2D and Pins libraries are
of a known working version. It will prevent the code from compiling
unless any mismatch is dealt with.

The code is setup to allow pin translation to work, though I have not
tested it yet.